Add BFS numpy
