+++ cayleypy/__init__.py
@@ -1,3 +1,4 @@
 from .cayley_graph import CayleyGraph
 from .graphs_lib import prepare_graph
 from .datasets import load_dataset
+from .bfs_special import bfs_numpy
