+++ cayleypy/cayley_graph_test.py
@@ -198,16 +198,7 @@ def test_generators_not_inverse_closed():
         graph.bfs()
 
 
-def test_bfs_numpy():
-    graph = prepare_graph("lrx", n=7)
-    assert graph.bfs_numpy() == load_dataset("lrx_cayley_growth")["7"]
 
-    graph = prepare_graph("top_spin", n=7)
-    assert graph.bfs_numpy() == load_dataset("top_spin_cayley_growth")["7"]
-
-    dest = "000000000111111111"
-    graph = CayleyGraph(prepare_graph("top_spin", n=18).generators, dest=dest)
-    assert graph.bfs_numpy() == load_dataset("top_spin_coset_growth")[dest]
 
 
 # Tests below compare growth function for small graphs with stored pre-computed results.
